backend/pyproject.toml
[project]
name = "backend"
version = "0.3.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aico-shared",
    "chromadb>=1.0.16",
    "duckdb>=1.3.2",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "libsql==0.1.8",
    "libsql-client>=0.3.1",
    "passlib[bcrypt]>=1.7.4",
    "platformdirs>=4.0.0",
    "protobuf>=6.32.0,<7.0.0",
    "psutil>=6.1.0",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "pyzmq>=27.0.1",
    "uvicorn>=0.35.0",
]

[tool.uv.sources]
aico-shared = { path = "../shared", editable = true }
